add photo item from api, fix genre, fix transition to DetailedItemFragment

+++ app/src/main/java/com/example/reelygoodmovies/ui/add_edit_movie/AddOrEditItemFragment.kt
@@ -186,7 +186,7 @@ class AddOrEditItemFragment : Fragment() {
         )
         setNumberPickers(movie)
         viewModel.setSelectedYear(movie.year)
-        binding.rbMovieRating.rating = movie.rate.toFloat()
+        binding.rbMovieRating.rating = movie.rate
         showGenres(movie)
         viewModel.setSelectedImageURI(movie.photo)
     }
